requirements.txt
Flask==2.2.5
scikit-learn==1.2.2
numpy==1.24.3
anthropic==0.25.0
python-dotenv==1.0.0
pdftotext==2.3.12
